backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-RESTX==1.3.0
PyPDF2==3.0.1
python-docx==1.1.0
pdfplumber==0.10.3
email-validator==2.1.0
openai==1.6.1
python-dotenv==1.0.0
requests==2.31.0
